backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
httpx==0.27.0
python-multipart==0.0.12
Pillow==10.4.0
ruff==0.6.0
pytest==8.3.0
pytest-asyncio==0.24.0
